Separate tests for Linear and MultiProc test.
